Add files via upload
The Slideshow was a more difficult then i thought. I first found 8 same size images that i could use then i made each one of the images into it own container and hid all of them except the first. then when the arrows are pressed it hides the existing image and displays the next. The difficult part was making it so that the arrows were postioned correctly as they kept showing on the right side of the image and wernt displaing. I had to add some CSS elements in order to move the arrows to the left and right side of the image.